+++ test/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+    "test.fixtures.pre_production.state_preprod_fg4_t4d_y1_x1",
+]
