+++ test/function/core/test_cell_area.py
